Add greater number comparison
This commit adds a comparison between two integer inputs, `number1` and `number2`, to determine which one is greater. It then displays the greater number along with the message "Greater number." This change enhances the program by providing a way to compare and identify the larger of the two input numbers.