Add Clothing and Shoes Recommendation
Implemented code to provide clothing and shoes recommendations based on the temperature (temp) and the time of day (dayTime). The program considers various temperature and time scenarios to suggest appropriate attire and footwear. It then generates a message to inform the user about what to wear. This update enhances the program's functionality to offer clothing and shoes recommendations.