Add solution for day of the week task
In this task, I've implemented a C# program to retrieve the name of a day of the week based on the user's input number (1-7). The solution uses an array of day names and a conditional statement to ensure that the input is within a valid range. If the input is valid, it displays the corresponding day's name; otherwise, it outputs "Invalid day!" to inform the user. This task marks the beginning of our exploration into arrays and their practical applications in C# programming.