Add code solution for rotating a list of integers
This code solution allows you to rotate a list of integers a specified number of times. The program takes the first number in the list and moves it to the end of the list, repeating this process according to the specified number of rotations. The rotated list is then displayed as the output.