Add code solution for finding the longest sequence of equal elements
This code solution is designed to find the longest sequence of equal elements in a list of integers. It iterates through the list, tracking the current length of the sequence and the maximum length found so far. If a longer sequence is encountered, the program updates the maximum length and the associated element. The final output is a list containing the element with the longest sequence repeated the corresponding number of times.