Add code solution for finding top integers in a list
This code solution is aimed at identifying and extracting the "top integers" from a given list of integers. An integer is considered a "top integer" if it is greater than all the integers to its right in the list. The code iterates through the list and, for each integer, checks if it is greater than or equal to all the integers to its right. If an integer meets this criterion, it is identified as a "top integer" and added to a separate list. The final output is a list containing all the "top integers" found in the original list.