Add EvenOddSubtraction class
Added the "EvenOddSubtraction" class, which includes the "FindDifference" method. This method takes an integer array as input and calculates the absolute difference between the sums of even and odd numbers in the array. This class provides a convenient utility for finding the difference between sums of even and odd integers, contributing to a well-rounded set of tools for working with arrays.