Add the Reverser class for string array reversal
In this commit, the "Reverser" class is introduced to the project. This class provides a method, "ReverseStrings," which takes an array of strings and returns a new array with the strings reversed. The reversal is done by transforming each string's characters in reverse order. This functionality will be useful for various string manipulation tasks within the application.