Add CenterPoint Class and Unit Tests
In this commit, I've added the `CenterPoint` class, which contains the `GetClosest` method. The `GetClosest` method takes coordinates of two points (x1, y1, x2, y2) and determines which point is closer to the origin (0, 0). I've also included a set of unit tests for this method to validate its functionality in different scenarios. These tests cover cases where one or both points have negative coordinates, as well as situations where both points are equidistant from the origin. The tests ensure the correct behavior of the `GetClosest` method.